Cast zlib malloc size to size_t before multiply.

+++ src/server/main.c
@@ -2152,7 +2152,7 @@ static void sv_hostname_changed(cvar_t *self)
 #if USE_ZLIB
 voidpf SV_zalloc(voidpf opaque, uInt items, uInt size)
 {
-    return SV_Malloc(items * size);
+    return SV_Malloc((size_t)items * size);
 }
 
 void SV_zfree(voidpf opaque, voidpf address)
